Create Tube Furnaces @ ENNE Synergy
Tube furnaces are often used in the heat processing steps of refining silicon for use in microchips and transistors due to their ability to safely reach temperatures of over 1000 oC using less power than that of typical ovens. The tube shape of the furnace allows for the introduction of steam and other diffusing substances using the process of heating, allowing for rapid oxidation or diffusion of dopants onto polished silicon. However, efficient tube furnaces often come with four figure plus price tags attached, and are often lab grade machines to withstand the repeated subjection to high temperatures. Despite this, tube furnaces are not particularly complex in construction, if the user's goal is simply to use the machine to safely achieve the high temperatures needed to create semiconductors. Although a homemade furnace would likely not be up to standard with an industrial tube furnace, the absence of luxuries such as ease of diffusion, control system regulation of temperature, and minimizing external factors means that, in an educational environment, forging semiconductors can be simplified to a high school level of understanding. Additionally, there is more freedom in the process as a result, allowing for instructors to slow down or explain certain processes to a greater degree, and for hobbyists to have more control over certain aspects in the process.